requirements_development.txt
flask
flask-cors
mysql-connector-python
python-dotenv
ruff